Fix build with CUDA
